requirements.txt
Flask==2.2.5
joblib==1.2.0
python-dotenv==1.0.0
scikit-learn==1.4.2
pandas==1.5.3
openpyxl>=3.1.5
appdirs==1.4.3
cached-property==1.3.1
certifi>=2023.7.22
chardet>=4.0.0
defusedxml==0.5.0
idna>=3.4
isodate==0.6.1
lxml>=4.6.5
pytz>=2021.3
requests>=2.28.2
requests-toolbelt==0.9.1
six>=1.16.0
urllib3>=1.26.14
zeep>=4.2.1
pycryptodome>=3.9.8
websocket-client>=1.6.1
cryptography>=3.4.8